Fixes for building with Visual C++ compilers. Tested with VC++ 6 (32-bit) and 2003 SDK compiler (64-bit).
C sources changed to conform to C89 (instead of C99).
makefile.vc and nmakehlp.c updated (too many changes to list)
Build expects Itcl to reside in a directory called itcl that is
the sibling of the itk directory.
nmake install also works but nmake winhelp (docs) does not
as it relies on an external program hrtf.exe which I do
not have.

Only verified with the demo itkedit script after modifying
the Iwidgets package to use Itk 4 instead of Itk 3.
The test scripts fail with errors that I believe are not
related to Windows.